% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pMAD_BM}
\alias{pMAD_BM}
\title{CDF of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval}
\usage{
pMAD_BM(q, summands = 100)
}
\arguments{
\item{q}{the quantity at which CDF will be evaluated. Currently accepts only a scalar}

\item{summands}{maximum number of terms to be evaluated in the infinite series (default=100)}
}
\value{
a scalar value
}
\description{
CDF of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval
}
