% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qMAD_BM}
\alias{qMAD_BM}
\title{Quantile function of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval}
\usage{
qMAD_BM(p)
}
\arguments{
\item{p}{the quantity at which the quantile function will be evaluated. Currently accepts only a scalar}
}
\value{
a scalar value
}
\description{
Quantile function of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval
}
