% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qMAD_BM_c}
\alias{qMAD_BM_c}
\title{Quantile function of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval, conditional on its terminal value}
\usage{
qMAD_BM_c(p, w1)
}
\arguments{
\item{p}{the quantity at which the quantile function will be evaluated. Currently accepts only a scalar}

\item{w1}{the terminal value}
}
\value{
a scalar value
}
\description{
Quantile function of the distribution of the maximum absolute deviation of Brownian motion in [0,1] interval, conditional on its terminal value
}
