% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulcalib-package.R
\docType{package}
\name{cumulcalib-package}
\alias{cumulcalib}
\alias{cumulcalib-package}
\title{cumulcalib: Cumulative calibration assessment}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msadatsafavi.github.io/cumulcalib/}
}

}
\author{
\strong{Maintainer}: Mohsen Sadatsafavi \email{Mohsen.sadatsafavi@ubc.ca} (\href{https://orcid.org/0000-0002-0419-7862}{ORCID})

}
\keyword{internal}
