% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulcalib-package.R
\docType{package}
\name{cumulcalib-package}
\alias{cumulcalib-package}
\alias{_PACKAGE}
\title{cumulcalib: Cumulative Calibration assessment}
\description{
This package implements the cumulative calibration error assessment for risk prediction models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msadatsafavi.github.io/cumulcalib/}
}

}
\author{
\strong{Maintainer}: Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} (\href{https://orcid.org/0000-0002-0419-7862}{ORCID})

}
\keyword{internal}
