% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulcalib-package.R
\docType{package}
\name{cumulcalib-package}
\alias{cumulcalib-package}
\alias{_PACKAGE}
\title{cumulcalib: Cumulative Calibration Assessment For Prediction Models}
\description{
Tools for visualization of, and inference on, the calibration of prediction models on the cumulative calibration domain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/resplab/cumulcalib}
}

}
\author{
\strong{Maintainer}: Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} (\href{https://orcid.org/0000-0002-0419-7862}{ORCID})

}
\keyword{internal}
