% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cumulcalibITE}
\alias{cumulcalibITE}
\title{Cumulative calibration assessment for individual treatment effect (ITE) predictions}
\usage{
cumulcalibITE(
  y,
  h,
  a,
  p = NULL,
  method = c("BB", "BM"),
  ordered = FALSE,
  n_sim = 0,
  aux = FALSE
)
}
\arguments{
\item{y}{vector of binary responses}

\item{h}{vector of predicted treatment benefits (the predicted reduction in outcome risk due to treatment).}

\item{a}{treatment indicator (1 if treated, 0 if control).}

\item{p}{optional vector of predicted baseline risks (the risk without treatment). If omitted (NULL), the marginal test is performed using observed event rates in the treated and control groups; if supplied, the conditional test is performed. Default is NULL.}

\item{method}{string with either BB (Brownian bridge test, default method), BM (Brownian motion test), BM2p (two-part BM test - experimental), BB1p (one-part BB test wit only the 'bridge' component). Multiple methods can be specified. The first one will be the 'main' method (e.g., when submitting the resulting object to plot()). Default is c("BB","BM")}

\item{ordered}{if TRUE, the data are already ordered based on ascending values of h. This is to speed up simulations.}

\item{n_sim}{if >0, indicates a simulation-based test is requested for inference.}

\item{aux}{if TRUE, auxiliary quantities (used internally and for diagnostics) are returned in the result. Default is FALSE.}
}
\value{
an objective of class cumulcalib that can be printed or plotted
}
\description{
This is the core function for performing cumulative calibration assessment for
models that predict individual treatment effects (treatment benefit).
}
\examples{
p <- rbeta(1000, 1, 2)
a <- rbinom(length(p), 1, 0.5)
h <- runif(length(p), 0, 0.05)
Y <- rbinom(length(p), 1, pmax(0, pmin(1, p - a * h)))
res <- cumulcalibITE(Y, h = h, a = a)
summary(res)
plot(res)
}
