% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pKolmogorov}
\alias{pKolmogorov}
\title{CDF of the Kolmogorov distribution}
\usage{
pKolmogorov(q, summands = ceiling(q * sqrt(72) + 3/2))
}
\arguments{
\item{q}{the quantity at which CDF will be evaluated. Currently accepts only a scalar}

\item{summands}{maximum number of terms to be evaluated in the infinite series (default=ceiling(q*sqrt(72)+3/2))}
}
\value{
a scalar value
}
\description{
CDF of the Kolmogorov distribution
}
