% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot.cumulcalib}
\alias{plot.cumulcalib}
\title{Generates cumulative calibration plots}
\usage{
\method{plot}{cumulcalib}(
  cumulcalib_obj,
  method = NULL,
  draw_stats = list(sigs = c(p1 = 0.95, p2 = 0.95), cols = c(p1 = "blue", p2 = "red")),
  x2axis = T,
  y2axis = T,
  ...
)
}
\arguments{
\item{cumulcalib_obj}{An object of class cumulcalib generated by cumulcalib()}

\item{method}{Which method to use. Options are BB (Brownian bridge test), BM (Brownian motion test), BB1p (1-part Brownian bridge test), and BM2p (2-part Brownian bridge test). If unspecified, returns the default method used in the cumulcalib() call}

\item{draw_stats}{a list with two elements: sigs=significance levels (a scalar if one-part test, a vector of 2 if two-part test) and cols=colors (a character string if one-part test, a vector of two-character strings if two-part test). Default is list(sigs=c(p1=0.95,p2=0.95), cols=c(p1='blue',p2='red'))}

\item{x2axis=T}{If true, draws a second x-axis (on top) showing predicted risks}

\item{y2axis=T}{If true, draws a second y-axis (on right) showing scaled partial sums}
}
\value{
None
}
\description{
Generates cumulative calibration plots
}
