% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.cumulcalib}
\alias{print.cumulcalib}
\title{Concise printing of a cumulcalib object}
\usage{
\method{print}{cumulcalib}(x, ...)
}
\arguments{
\item{x}{An object of class cumulcalib generated by cumulcalib() or cumulcalibITE()}

\item{...}{Not used}
}
\value{
The input object, invisibly.
}
\description{
Prints a one-screen overview of a calibration assessment returned by
\code{\link[=cumulcalib]{cumulcalib()}} or \code{\link[=cumulcalibITE]{cumulcalibITE()}}. For a fuller breakdown use \code{\link[=summary]{summary()}}.
}
