% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.cumulcalib}
\alias{print.cumulcalib}
\title{Summarizes a cumul_calib object}
\usage{
\method{print}{cumulcalib}(cumulcalib_obj, method = NULL)
}
\arguments{
\item{cumulcalib_obj}{An object of class cumulcalib generated by cumulcalib()}

\item{method}{Which method to use. Options are BB (Brownian bridge test), BM (Brownian motion test), BB1p (1-part Brownian bridge test), and BM2p (2-part Brownian bridge test). If unspecified, returns the default method used in the cumulcalib() call}
}
\value{
None
}
\description{
Summarizes a cumul_calib object
}
