% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.summary.cumulcalib}
\alias{print.summary.cumulcalib}
\title{Print a cumulcalib summary}
\usage{
\method{print}{summary.cumulcalib}(x, ...)
}
\arguments{
\item{x}{An object of class summary.cumulcalib generated by summary()}

\item{...}{Not used}
}
\value{
The input object, invisibly.
}
\description{
Print a cumulcalib summary
}
