% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{qKolmogorov}
\alias{qKolmogorov}
\title{Quantile function of the Kolmogorov distribution}
\usage{
qKolmogorov(p)
}
\arguments{
\item{p}{the quantity at which the quantile function will be evaluated. Currently accepts only a scalar}
}
\value{
a scalar value
}
\description{
Quantile function of the Kolmogorov distribution
}
